Switched BN (conditional and non) to the one from tfgan. this seems to solve the eval difference between TPU and GPU. the old BN code currently exists in op.py but is not used.
replacing the cond-BN in the generator required another minor change-
this version of BN currently requires one-hot class encoding (the old one took int). argmax is taken in the bn code, so currently it goes int to one-hot and back (non-ideal). consider changing the bn code and then move back to providing int.

PiperOrigin-RevId: 269172811
Change-Id: Iecdd06131ded03936c134e1c8243c65dde595a07